# be/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PlanWise API Documentation",
        "description": "API Documentation for PlanWise application - a smart daily planning application with AI features.",
        "contact": {
            "name": "PlanWise Team",
            "email": "contact@planwise.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://springdoc.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development"
        }
    ],
    "security": [
        {
            "bearer": []
        }
    ],
    "paths": {
...
        "/api/v1/focus/daily/notes": {
            "put": {
                "tags": [
                    "focus-controller"
                ],
                "operationId": "updateDailyNotes",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-06-05"
                        },
                        "example": "2026-06-05"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DailyFocusDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/focus/sessions": {
            "get": {
                "tags": [
                    "focus-controller"
                ],
                "operationId": "getFocusSessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FocusSessionDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "focus-controller"
                ],
                "operationId": "createFocusSession",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFocusSessionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FocusSessionDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/focus/notes": {
            "get": {
                "tags": [
                    "focus-controller"
                ],
                "operationId": "getQuickNotes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuickNoteDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "focus-controller"
                ],
                "operationId": "createQuickNote",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateQuickNoteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuickNoteDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/focus/daily/{date}/top-tasks/{taskId}": {
            "post": {
                "tags": [
                    "focus-controller"
                ],
                "operationId": "addTopTask",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-06-05"
                        },
                        "example": "2026-06-05"
                    },
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DailyFocusDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "focus-controller"
                ],
                "operationId": "removeTopTask",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-06-05"
                        },
                        "example": "2026-06-05"
                    },
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DailyFocusDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/focus/sessions/{sessionId}": {
            "patch": {
                "tags": [
                    "focus-controller"
                ],
                "operationId": "updateFocusSession",
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FocusSessionDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/focus/daily": {
            "get": {
                "tags": [
                    "focus-controller"
                ],
                "operationId": "getDailyFocus",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-06-05"
                        },
                        "example": "2026-06-05"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DailyFocusDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/focus/notes/{noteId}": {
            "delete": {
                "tags": [
                    "focus-controller"
                ],
                "operationId": "deleteQuickNote",
                "parameters": [
                    {
                        "name": "noteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "DailyFocusDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "focusDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-05"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "topTaskIds": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        }
                    },
                    "quickNotes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QuickNoteDto"
                        }
                    },
                    "focusSessions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FocusSessionDto"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    }
                }
            },
            "FocusSessionDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "startTime": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "duration": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sessionType": {
                        "type": "string",
                        "example": "POMODORO"
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "endTime": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "taskId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "taskTitle": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    }
                }
            },
            "QuickNoteDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "content": {
                        "type": "string"
                    },
                    "noteType": {
                        "type": "string"
                    },
                    "mediaUrl": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    }
                }
            },
...
            "CreateFocusSessionRequest": {
                "type": "object",
                "properties": {
                    "startTime": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "duration": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sessionType": {
                        "type": "string",
                        "example": "POMODORO"
                    },
                    "taskId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "notes": {
                        "type": "string"
                    }
                },
                "required": [
                    "startTime"
                ]
            },
            "CreateQuickNoteRequest": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "string"
                    },
                    "noteType": {
                        "type": "string"
                    },
                    "mediaUrl": {
                        "type": "string"
                    }
                },
                "required": [
                    "content"
                ]
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Enter your Bearer JWT token to access protected endpoints.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}